requirements.txt
aiogram==3.6.0
matplotlib>=3.5.0
pandas>=1.3.0
openpyxl>=3.0.0
pulp>=2.6.0
python-docx>=0.8.11
python-dotenv>=0.19.0
reportlab>=4.0.0

# Новые зависимости для системы расчёта ПБ ЖБК СТАРТ
# (все остальные уже были в проекте)
